requirements.txt
discord
feedparser
dotenv
pyyaml

